Extract performance confirmation into hook
